feat: keychain, connection profiles, S3 mount dialog, quick connect
Keychain:
- System keychain integration via keyring crate (macOS Keychain,
  Linux Secret Service). Commands: keychain_get/set/delete.

Connection profiles:
- Persisted in ~/.config/newt/connections.toml with CRUD commands.
- Secrets stored in system keychain under "connection:<id>".
- Types: S3 (region, bucket, endpoint, credential mode, profile,
  role ARN), SFTP (host), Remote (host).
- connect_profile command handles everything server-side: reads
  profile, loads secrets, dispatches to mount/connect.

S3 mount dialog:
- Full credential configuration: Default / AWS Profile / IAM User /
  Assume Role with endpoint URL override and scoped bucket support.
- Save as connection profile with auto-generated name.

Quick Connect (Ctrl+R):
- Command palette UI listing saved connections with fuzzy search,
  keyboard nav, Delete to remove. Frontend just calls connect_profile.

Other changes:
- Save as profile in SFTP and Remote connect dialogs.
- DisplayPathMatch with Ord priority (Exact > Generic, mount order
  tiebreaker).
- Copy/move dialog: read-only destination, no ambiguous path
  resolution.
- Editor: Escape closes with dirty prompt.
- Fix i128 timestamp overflow panic (clamp to i64).
- Refresh command unbound by default (was Mod+R, clashed with
  Ctrl+R quick connect).

+++ libs/newt-common/src/file_reader.rs
@@ -23,9 +23,9 @@ pub struct FileDetails {
     pub user: Option<UserGroup>,
     pub group: Option<UserGroup>,
     pub mode: Option<Mode>,
-    pub modified: Option<i128>,
-    pub accessed: Option<i128>,
-    pub created: Option<i128>,
+    pub modified: Option<i64>,
+    pub accessed: Option<i64>,
+    pub created: Option<i64>,
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
